- WARN  | [iOS]  MessageDisplayKit/MessageDisplayKit/Classes/Controllers/XHMessageTableViewController/XHMessageTableViewController.m:710:1: warning: conflicting return type in implementation of 'supportedInterfaceOrientations': 'UIInterfaceOrientationMask' (aka 'enum UIInterfaceOrientationMask') vs 'NSUInteger' (aka 'unsigned int') [-Wmismatched-return-types]

+++ MessageDisplayKit/Classes/Controllers/XHMessageTableViewController/XHMessageTableViewController.m
@@ -708,7 +708,7 @@ static CGPoint  delayOffset = {0.0};
 }
 
 - (NSUInteger)supportedInterfaceOrientations {
-    return UIInterfaceOrientationMaskPortrait;
+    return (NSUInteger)UIInterfaceOrientationMaskPortrait;
 }
 
 - (UIInterfaceOrientation)preferredInterfaceOrientationForPresentation {
